refactor(leases): drain long/short form scripts into composables
Closes the remaining #185 scope. LongForm.vue (585) and ShortForm.vue
(561) held near-duplicate reactive-validation, quote, and submit logic
that pushed both SFCs well over the 400-LOC budget. Move the identical
infrastructure (stores, both watchers, shared computeds, calculate,
openLease, onOpenLease) into a position-parameterized useLeaseForm; the
genuinely divergent list-building and leaseTicker/protocol derivation
stay in per-form useLongLeaseForm / useShortLeaseForm behind a strategy,
so the Long-vs-Short asymmetry (Short stable via resolveShortLeaseTicker)
remains visible. Behaviour-preserving: the reactive balance-first bail,
classifyError classification, and the ShortForm ticker contract are
unchanged, and the LongForm/ShortForm characterization tests pass
untouched. Each SFC is now template + one destructure (~205 LOC).
